Issue #112: refactored all packages to name com.macbitsgoa.ard
