+++ app/src/main/java/com/macbitsgoa/ard/fragments/ChatFragment.java
@@ -1,0 +1,95 @@
+package com.macbitsgoa.ard.fragments;
+
+import android.content.Context;
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.macbitsgoa.ard.R;
+import com.macbitsgoa.ard.interfaces.ChatFragmentListener;
+import com.macbitsgoa.ard.utils.AHC;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * Activities that contain this fragment can implement the
+ * {@link ChatFragmentListener} interface
+ * to handle interaction events.
+ * Use the {@link ChatFragment#newInstance} factory method to
+ * create an instance of this fragment.
+ *
+ * @author Vikramaditya Kukreja
+ */
+public class ChatFragment extends Fragment {
+
+    /**
+     * Tag for this fragment.
+     */
+    private final String TAG = AHC.TAG + ".fragments." + ChatFragment.class.getSimpleName();
+
+    /**
+     * Fragment title to be used.
+     */
+    private String fragmentTitle;
+
+    /**
+     * Used to communicate with activity.
+     */
+    private ChatFragmentListener mListener;
+
+    public ChatFragment() {
+        // Required empty public constructor
+    }
+
+    /**
+     * Use this factory method to create a new instance of
+     * this fragment using the provided parameters.
+     *
+     * @param fragmentTitle Title for the fragment.
+     * @return A new instance of fragment ChatFragment.
+     */
+    public static ChatFragment newInstance(final String fragmentTitle) {
+        final ChatFragment fragment = new ChatFragment();
+        final Bundle args = new Bundle();
+        args.putString(AHC.FRAGMENT_TITLE_KEY, fragmentTitle);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    @Override
+    public void onCreate(final Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        if (getArguments() != null) {
+            fragmentTitle = getArguments().getString(AHC.FRAGMENT_TITLE_KEY);
+        }
+    }
+
+    @Override
+    public View onCreateView(final LayoutInflater inflater, final ViewGroup container,
+                             final Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        mListener.updateChatFragment();
+        Log.d(TAG, fragmentTitle);
+        return inflater.inflate(R.layout.fragment_chat, container, false);
+    }
+
+    @Override
+    public void onAttach(final Context context) {
+        super.onAttach(context);
+        if (context instanceof ChatFragmentListener) {
+            mListener = (ChatFragmentListener) context;
+        } else {
+            throw new RuntimeException(context.toString()
+                    + " must implement OnFragmentInteractionListener");
+        }
+    }
+
+    @Override
+    public void onDetach() {
+        super.onDetach();
+        mListener = null;
+    }
+
+}
