+++ app/src/main/java/com/macbitsgoa/ard/interfaces/ChatFragmentListener.java
@@ -1,0 +1,9 @@
+package com.macbitsgoa.ard.interfaces;
+
+/**
+ * Listener for Chat fragment.
+ * @author Vikramaditya Kukreja
+ */
+public interface ChatFragmentListener {
+    void updateChatFragment();
+}
