+++ app/src/main/java/com/macbitsgoa/ard/interfaces/FaqFragmentListener.java
@@ -1,0 +1,9 @@
+package com.macbitsgoa.ard.interfaces;
+
+/**
+ * Listener for Faq fragment.
+ * @author Vikramaditya Kukreja
+ */
+public interface FaqFragmentListener {
+    void updateFaqFragment();
+}
