+++ app/src/main/java/com/macbitsgoa/ard/interfaces/HomeFragmentListener.java
@@ -1,0 +1,9 @@
+package com.macbitsgoa.ard.interfaces;
+
+/**
+ * Listener for Home fragment.
+ * @author Vikramaditya Kukreja
+ */
+public interface HomeFragmentListener {
+    void updateHomeFragment();
+}
