+++ app/src/main/java/com/macbitsgoa/ard/utils/BrowserUtil.java
@@ -1,0 +1,25 @@
+package com.macbitsgoa.ard.utils;
+
+/**
+ * Constant value definitions for {@link Browser} class.
+ * @author Rushikesh Jogdand
+ */
+
+@SuppressWarnings("WeakerAccess")
+public class BrowserUtil {
+
+    /**
+     * Height of action button (must be 24dp according to docs).
+     */
+    public static final int ACTION_BUTTON_HEIGHT = 24;
+
+    /**
+     * Height of close button (must be 24dp according to docs).
+     */
+    public static final int CLOSE_BUTTON_HEIGHT = 24;
+
+    /**
+     * Width of close button (must be 24dp according to docs).
+     */
+    public static final int CLOSE_BUTTON_WIDTH = 24;
+}
