--- app/src/main/java/in/ac/bits_pilani/goa/ard/interfaces/ChatFragmentListener.java
@@ -1,9 +1,0 @@
-package in.ac.bits_pilani.goa.ard.interfaces;
-
-/**
- * Listener for Chat fragment.
- * @author Vikramaditya Kukreja
- */
-public interface ChatFragmentListener {
-    void updateChatFragment();
-}
