--- app/src/main/java/in/ac/bits_pilani/goa/ard/interfaces/FaqFragmentListener.java
@@ -1,9 +1,0 @@
-package in.ac.bits_pilani.goa.ard.interfaces;
-
-/**
- * Listener for Faq fragment.
- * @author Vikramaditya Kukreja
- */
-public interface FaqFragmentListener {
-    void updateFaqFragment();
-}
