--- app/src/main/java/in/ac/bits_pilani/goa/ard/interfaces/HomeFragmentListener.java
@@ -1,9 +1,0 @@
-package in.ac.bits_pilani.goa.ard.interfaces;
-
-/**
- * Listener for Home fragment.
- * @author Vikramaditya Kukreja
- */
-public interface HomeFragmentListener {
-    void updateHomeFragment();
-}
