--- app/src/main/java/in/ac/bits_pilani/goa/ard/utils/BrowserUtil.java
@@ -1,25 +1,0 @@
-package in.ac.bits_pilani.goa.ard.utils;
-
-/**
- * Constant value definitions for {@link Browser} class.
- * @author Rushikesh Jogdand
- */
-
-@SuppressWarnings("WeakerAccess")
-public class BrowserUtil {
-
-    /**
-     * Height of action button (must be 24dp according to docs).
-     */
-    public static final int ACTION_BUTTON_HEIGHT = 24;
-
-    /**
-     * Height of close button (must be 24dp according to docs).
-     */
-    public static final int CLOSE_BUTTON_HEIGHT = 24;
-
-    /**
-     * Width of close button (must be 24dp according to docs).
-     */
-    public static final int CLOSE_BUTTON_WIDTH = 24;
-}
