+++ app/src/main/res/layout/fragment_chat.xml
@@ -3,7 +3,7 @@
     android:id="@+id/fragment_chat_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="in.ac.bits_pilani.goa.ard.fragments.ChatFragment">
+    tools:context="com.macbitsgoa.ard.fragments.ChatFragment">
 
     <View
         android:layout_width="match_parent"
