+++ app/src/main/res/layout/fragment_faq.xml
@@ -3,7 +3,7 @@
     android:id="@+id/fragment_faq_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="in.ac.bits_pilani.goa.ard.fragments.FaqFragment">
+    tools:context="com.macbitsgoa.ard.fragments.FaqFragment">
 
     <include layout="@layout/fragment_faq_content" />
 </FrameLayout>
