--- app/src/test/java/in/ac/bits_pilani/goa/ard/utils/AhcTest.java
@@ -1,22 +1,0 @@
-package in.ac.bits_pilani.goa.ard.utils;
-
-import org.junit.Before;
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotEquals;
-
-/**
- * Tests for AHC.
- * @author vikramaditya
- */
-
-public class AhcTest {
-
-
-    @Test
-    public void testClassName() throws Exception{
-        String expected = "AHC";
-        assertEquals("Class name is wrong", expected, AHC.class.getSimpleName());
-    }
-}
